* Changed PacketHandler imports to machines submodule   * Moved "registerMessage" to PacketHandler.java
