+++ src/main/java/crazypants/enderio/machines/machine/obelisk/weather/GuiWeatherObelisk.java
@@ -15,9 +15,9 @@ import crazypants.enderio.base.gui.IconEIO;
 import crazypants.enderio.base.lang.LangFluid;
 import crazypants.enderio.base.machine.gui.GuiInventoryMachineBase;
 import crazypants.enderio.base.machine.gui.PowerBar;
-import crazypants.enderio.base.network.PacketHandler;
 import crazypants.enderio.machines.lang.Lang;
 import crazypants.enderio.machines.machine.obelisk.weather.TileWeatherObelisk.WeatherTask;
+import crazypants.enderio.machines.network.PacketHandler;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.entity.player.InventoryPlayer;
