+++ src/main/java/crazypants/enderio/machines/machine/tank/GuiTank.java
@@ -12,7 +12,7 @@ import com.enderio.core.client.render.RenderUtil;
 
 import crazypants.enderio.base.lang.LangFluid;
 import crazypants.enderio.base.machine.gui.GuiMachineBase;
-import crazypants.enderio.base.network.PacketHandler;
+import crazypants.enderio.machines.network.PacketHandler;
 import crazypants.enderio.machines.lang.Lang;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.renderer.GlStateManager;
